backend/pyproject.toml
[tool.poetry]
name = "architect-backend"
version = "0.1.0"
description = "AI-powered orchestration platform backend using LangGraph"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
# LangGraph and LangChain
langgraph = "^0.0.20"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
langchain-community = "^0.0.13"
# LLM Providers
google-generativeai = "^0.3.2"
langchain-google-genai = "^0.0.5"
groq = "^0.4.1"
# Vector Database
qdrant-client = "^1.7.0"
# PostgreSQL
supabase = "^2.3.4"
asyncpg = "^0.29.0"
# Web Scraping
beautifulsoup4 = "^4.12.3"
httpx = "^0.26.0"
aiohttp = "^3.9.1"
# WebSocket support
python-socketio = "^5.11.0"
# Utilities
tenacity = "^8.2.3"
tiktoken = "^0.5.2"
python-multipart = "^0.0.6"
websockets = "13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]